requirements.txt
certifi==2024.2.2
cloudinary==1.39.1
pyotp==2.9.0
pypng==0.20220715.0
PyQRCode==1.2.1
qrcode==7.4.2
six==1.16.0
typing_extensions==4.10.0
urllib3==2.2.1
